+++ rest-api/pom.xml
@@ -29,21 +29,9 @@
 		</dependency>
 
 		<dependency>
-			<groupId>io.fabric8</groupId>
-			<artifactId>kubernetes-model</artifactId>
-			<version>5.0.1</version>
-		</dependency>
-
-		<dependency>
-			<groupId>io.fabric8</groupId>
-			<artifactId>openshift-client</artifactId>
-			<version>5.0.1</version>
-		</dependency>
-
-		<dependency>
-			<groupId>io.fabric8</groupId>
-			<artifactId>kubernetes-client</artifactId>
-			<version>5.0.1</version>
+			<groupId>io.kubernetes</groupId>
+			<artifactId>client-java</artifactId>
+			<version>10.0.0</version>
 		</dependency>
 
 		<dependency>
