+++ rest-api/src/main/java/uk/gov/gchq/gaffer/Graph.java
@@ -1,4 +1,4 @@
-package com.example.springboottest;
+package uk.gov.gchq.gaffer;
 
 public class Graph {
