+++ rest-api/src/main/java/uk/gov/gchq/gaffer/JwtRequest.java
@@ -1,5 +1,4 @@
-package com.example.springboottest;
-
+package uk.gov.gchq.gaffer;
 import java.io.Serializable;
 
 public class JwtRequest implements Serializable {
