+++ rest-api/src/main/java/uk/gov/gchq/gaffer/JwtUserDetailsService.java
@@ -1,5 +1,4 @@
-package com.example.springboottest;
-
+package uk.gov.gchq.gaffer;
 import java.util.ArrayList;
 
 import org.springframework.security.core.userdetails.User;
