+++ rest-api/src/main/java/uk/gov/gchq/gaffer/WebSecurityConfig.java
@@ -1,5 +1,4 @@
-package com.example.springboottest;
-
+package uk.gov.gchq.gaffer;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
