+++ rest-api/src/test/java/uk/gov/gchq/gaffer/SpringbootTestApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.SpringbootTest;
+package uk.gov.gchq.gaffer;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
